+++ verik-compiler/regression/02-module-instantiation.kt
@@ -20,7 +20,7 @@
 
 import io.verik.core.*
 
-@SimTop
+@EntryPoint
 object M0 : Module() {
 
     @Make