+++ CHANGELOG.md
@@ -16,6 +16,7 @@
 - Separately specialize nested classes.
 - Support type parameterized imported classes.
 - Remove support for constructor overloading.
+- Replace `SimTop` and `SynthTop` with `EntryPoint`.
 
 ## [0.1.13]
 ### Importer